VBR quality is now a float parameter
git-svn-id: http://svn.xiph.org/trunk/speex@4076 0101bb08-14d6-0310-b084-bc0e0c8e3800

+++ libspeex/nb_celp.c
@@ -1241,10 +1241,10 @@ void nb_encoder_ctl(void *state, int request, void *ptr)
       (*(int*)ptr) = st->vbr_enabled;
       break;
    case SPEEX_SET_VBR_QUALITY:
-      st->vbr_quality = (*(int*)ptr);
+      st->vbr_quality = (*(float*)ptr);
       break;
    case SPEEX_GET_VBR_QUALITY:
-      (*(int*)ptr) = st->vbr_quality;
+      (*(float*)ptr) = st->vbr_quality;
       break;
    case SPEEX_SET_QUALITY:
       {
